+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support nested callbacks inside array arguments, for instance in array_modify, array_create
 - Support array_concat
+- add `ProcessGraphUnflattener` and `PGNodeGraphUnflattener` to unflatten a flat dict representation of a process
+  graph to a `PGNode` graph (EP-3609)
+- Add `Connection.datacube_from_flat_graph` and `Connection.datacube_from_json` to construct a `DataCube`
+  from flat process graph representation (e.g. JSON file or JSON URL) (EP-3609)
 
 ### Changed